--- misk-web/web/packages/@misk/core/src/ducks/simpleDucksUtilities.ts
@@ -1,131 +1,0 @@
-/**
- * simple*Ducks libraries common code
- */
-
-/**
- * Handler Functions
- * Reduce the legwork of parsing `onChange`, `onClick` and other events in Buttons, Inputs, Toggles...etc to call your handling function
- *
- * Old Way: Manually declare inline arrow funtions consuming the input events
- * ```
- * <InputGroup onChange={(event: any) => (props.simpleFormInput("FormInputTag", event.target.value))}
- * <NumberInput onChange={({valueAsNumber: number, valueAsString: string}) => (props.simpleFormNumber("FormNumberTag", valueAsNumber, valueAsString))}
- * ```
- *
- * New way: Use on*FnCall to wrap and implicitly pass into handling functions the input events
- * ```
- * <InputGroup onChange={onChangeFnCall(props.simpleFormInput, ["FormInputTag"])}
- * <NumberInput onChange={onChangeNumberFnCall(props.simpleFormNumber, ["FormNumberTag"])}
- * ```
- */
-
-/**
- *
- * @param callFn: function to be called
- * @param args: arguments to be passed into the callFn
- *
- * ```
- * <Button onClick={onClickFnCall(props.simpleNetworkPut, ["PutTag", { ...requestBody }])}
- * ```
- */
-export const onClickFnCall = (callFn: any, args: any[]) => (event: any) => {
-  callFn(...args)
-}
-
-/**
- *
- * @param callFn: function to be called
- * @param args: arguments to be passed into the callFn
- *
- * ```
- * <InputGroup onChange={onChangeFnCall(props.simpleFormInput, ["FormInputTag"])}
- * ```
- */
-export const onChangeFnCall = (callFn: any, args: any[]) => (event: any) => {
-  callFn(...args, event.target.value)
-}
-
-/**
- *
- * @param callFn: function to be called
- * @param args: arguments to be passed into the callFn
- *
- * ```
- * <Checkbox onChange={onChangeToggleFnCall(props.simpleFormToggle, ["FormToggleTag", simpleFormState])}
- * ```
- */
-export const onChangeToggleFnCall = (callFn: any, args: any[]) => (
-  event: any
-) => {
-  callFn(...args, event.target.value)
-}
-
-/**
- *
- * @param callFn: function to be called
- * @param args: arguments to be passed into the callFn
- *
- * ```
- * <NumberInput onChange={onChangeNumberFnCall(props.simpleFormNumber, ["FormNumberTag"])}
- * ```
- */
-export const onChangeNumberFnCall = (callFn: any, args: any[]) => (
-  valueAsNumber: number,
-  valueAsString: string
-) => {
-  callFn(...args, valueAsNumber, valueAsString)
-}
-
-/**
- *
- * @param callFn: function to be called
- * @param args: arguments to be passed into the callFn
- *
- * ```
- * <TagInput onChange={onChangeTagFnCall(props.simpleFormInput, ["FormTagsTag"])}
- * ```
- */
-export const onChangeTagFnCall = (callFn: any, args: any[]) => (
-  values: string[]
-) => {
-  callFn(...args, values)
-}
-
-/**
- * Utilities
- */
-
-/**
- * @param oldState input from the event.target.value of a button (string)
- *                 or the oldState from Redux store (boolean)
- */
-export const booleanToggle = (oldState: string | boolean) => {
-  if (oldState === true || oldState === "on") {
-    return false
-  } else {
-    return true
-  }
-}
-
-/**
- * @param payload `action.payload` from Redux
- * Assumes that the first non-order safe key accessed is the data
- * Only use when action.payload has a single key (ie. the tag with all metadata inside)
- * Otherwise, unpredictable key selection
- */
-export const getPayloadTag = <T = any>(payload: { [tag: string]: T }) => {
-  return payload[Object.keys(payload)[0]]
-}
-
-/**
- *
- * @param json likely JSON input as a string
- * @returns JSON or string if JSON.parse fails
- */
-export const jsonOrString = (json: string) => {
-  try {
-    return JSON.parse(json)
-  } catch (e) {
-    return json
-  }
-}
