+++ misk-web/web/packages/@misk/core/src/index.ts
@@ -1,6 +1,5 @@
 export * from "./builders"
 export * from "./components"
 export * from "./containers"
-export * from "./ducks"
 export * from "./features"
 export * from "./utilities"
