+++ misk-web/web/packages/@misk/simpleredux/.npmignore
@@ -1,0 +1,5 @@
+src
+test
+coverage
+node_modules
+.npmrc
