+++ misk-web/web/packages/@misk/simpleredux/README.md
@@ -1,0 +1,30 @@
+## Misk Simple Redux
+
+![](https://raw.githubusercontent.com/square/misk/master/misk.png)
+Using Redux can be simple.
+
+`@misk/SimpleRedux` includes all the Redux parts necessary for forms, network requests, and other interactive components. Don't write Redux boilerplate for basic interactivity again.
+
+## Getting Started
+
+```bash
+$ npm install @misk/simpleredux
+```
+
+## SimpleForm
+
+A standardized set of form and input handler Redux-Sagas parts (actions, dispatcher, handlers, sagas, reducers, state interface)
+
+[Example Code](https://github.com/square/misk-web/blob/master/examples/tabs/palette-exemplar/src/containers/SampleFormContainer.tsx)
+
+## SimpleNetwork
+
+A standardized set of Axios based request Redux-Sagas parts (actions, dispatcher, handlers, sagas, reducers, state interface)
+
+[Example Code](https://github.com/square/misk-web/blob/master/examples/tabs/palette-exemplar/src/containers/SampleNetworkContainer.tsx)
+
+[Understanding Redux](https://github.com/square/misk-web/blob/master/misk-web/web/packages/%40misk/simpleredux/REDUX.md)
+
+## [Releasing](https://github.com/square/misk-web/blob/master/RELEASING.md)
+
+## [Changelog (and Breaking Changes)](https://github.com/square/misk-web/blob/master/CHANGELOG.md)
