+++ misk-web/web/packages/@misk/simpleredux/package.json
@@ -1,0 +1,117 @@
+{
+  "name": "@misk/simpleredux",
+  "version": "0.1.4-3",
+  "license": "SEE LICENSE IN https://github.com/square/misk-web",
+  "description": "Misk-Web SimpleRedux",
+  "author": "Square/Misk Authors (https://github.com/square/misk-web/graphs/contributors)",
+  "main": "lib/web/@misk/simpleredux/simpleredux.js",
+  "module": "lib/web/@misk/simpleredux/simpleredux.js",
+  "typings": "lib/web/@misk/simpleredux/index.d.ts",
+  "engines": {
+    "node": ">=7.8.0",
+    "yarn": "YARN NO LONGER USED - use npm instead."
+  },
+  "files": [
+    "lib"
+  ],
+  "repository": {
+    "type": "git",
+    "url": "git@github.com:square/misk-web.git"
+  },
+  "scripts": {
+    "build": "npm run-script lint && npm run-script lib",
+    "clean": "rm -rf lib",
+    "dev": "npm run-script build; npm run-script updateDevCache",
+    "ci-build": "npm run-script clean && npm run-script lib",
+    "lib": "webpack",
+    "lint": "prettier --write --config package.json \".{/src/**/,/}*.{md,css,sass,less,json,js,jsx,ts,tsx}\"",
+    "gradle": "cd ../../../../../ && ./gradlew web -Pcmd='-ca' -Ptabs='@misk/simpleredux'",
+    "prepare": "npm run-script ci-build",
+    "reinstall": "rm -rf node_modules && npm install",
+    "start": "npm run-script lint && cross-env NODE_ENV=development webpack-dev-server",
+    "test": "echo no tests",
+    "updateDevCache": "cd ..; ./updateLocalCache.sh"
+  },
+  "dependencies": {
+    "@misk/common": "^0.1.4-3",
+    "@blueprintjs/core": "^3.9.0",
+    "@blueprintjs/datetime": "^3.4.0",
+    "@blueprintjs/icons": "^3.3.0",
+    "@blueprintjs/select": "^3.3.0",
+    "@blueprintjs/table": "^3.3.0",
+    "@blueprintjs/timezone": "^3.1.0",
+    "axios": "^0.18.0",
+    "connected-react-router": "^5.0.1",
+    "dayjs": "^1.7.7",
+    "history": "^4.7.2",
+    "immutable": "^4.0.0-rc.12",
+    "lodash-es": "^4.17.11",
+    "react": "^16.6.3",
+    "react-dom": "^16.6.3",
+    "react-helmet": "^5.2.0",
+    "react-hot-loader": "^4.3.12",
+    "react-redux": "^5.1.1",
+    "react-router": "^4.3.1",
+    "react-router-dom": "^4.3.1",
+    "react-transition-group": "^2.5.0",
+    "redux": "^4.0.1",
+    "redux-saga": "^0.16.2",
+    "reselect": "^4.0.0",
+    "re-reselect": "2.2.0",
+    "skeleton-css": "^2.0.4",
+    "styled-components": "^4.1.2"
+  },
+  "devDependencies": {
+    "@misk/dev": "^0.1.4-3",
+    "@misk/tslint": "^0.1.4-3",
+    "@types/history": "^4.7.2",
+    "@types/lodash-es": "^4.17.1",
+    "@types/node": "^10.12.18",
+    "@types/react": "^16.7.18",
+    "@types/react-dom": "^16.0.11",
+    "@types/react-helmet": "^5.0.8",
+    "@types/react-hot-loader": "^4.1.0",
+    "@types/react-redux": "^6.0.12",
+    "@types/react-router": "^4.4.3",
+    "@types/react-router-dom": "^4.3.1",
+    "@types/styled-components": "^4.1.4",
+    "@types/webpack": "^4.4.22",
+    "@types/webpack-env": "^1.13.6",
+    "copy-webpack-plugin": "^4.6.0",
+    "cross-env": "^5.2.0",
+    "css-loader": "^2.1.0",
+    "file-loader": "^3.0.1",
+    "html-webpack-harddisk-plugin": "^1.0.1",
+    "html-webpack-plugin": "^3.2.0",
+    "mini-css-extract-plugin": "^0.5.0",
+    "optimize-css-assets-webpack-plugin": "^5.0.1",
+    "prettier": "^1.15.3",
+    "prop-types": "^15.6.2",
+    "sass-loader": "^7.1.0",
+    "source-map-loader": "^0.2.4",
+    "style-loader": "^0.23.1",
+    "ts-loader": "^5.3.3",
+    "tslib": "^1.9.3",
+    "typescript": "^3.2.2",
+    "typescript-plugin-styled-components": "^1.0.0",
+    "webpack": "^4.28.3",
+    "webpack-cli": "^3.2.1",
+    "webpack-bundle-analyzer": "^3.0.3",
+    "webpack-dev-server": "^3.1.14",
+    "webpack-merge": "^4.2.1"
+  },
+  "prettier": {
+    "arrowParens": "avoid",
+    "bracketSpacing": true,
+    "jsxBracketSameLine": false,
+    "printWidth": 80,
+    "semi": false,
+    "singleQuote": false,
+    "tabWidth": 2,
+    "trailingComma": "none",
+    "useTabs": false
+  },
+  "miskTab": {
+    "version": "0.1.1"
+  }
+}
