+++ misk-web/web/packages/@misk/simpleredux/prettier.config.js
@@ -1,0 +1,2 @@
+const { createPrettierConfig } = require("@misk/dev")
+module.exports = createPrettierConfig()
