+++ misk-web/web/packages/@misk/simpleredux/tslint.json
@@ -1,0 +1,3 @@
+{
+  "extends": "@misk/tslint"
+}
