+++ misk-web/web/packages/rush.json
@@ -273,6 +273,11 @@
       "versionPolicyName": "MiskWeb"
     },
     {
+      "packageName": "@misk/simpleredux",
+      "projectFolder": "@misk/simpleredux",
+      "versionPolicyName": "MiskWeb"
+    },
+    {
       "packageName": "@misk/tslint",
       "projectFolder": "@misk/tslint",
       "versionPolicyName": "MiskWeb"
